move function prepare_output_cmd, add test video_audio_output
