+++ tests/src/lib_utils.rs
@@ -50,66 +50,3 @@ fn test_delta() {
 
     assert!(delta < 2.0);
 }
-
-#[test]
-fn test_prepare_output_cmd() {
-    let enc_prefix = vec_strings![
-        "-hide_banner",
-        "-nostats",
-        "-v",
-        "level+error",
-        "-re",
-        "-i",
-        "pipe:0"
-    ];
-    let filter = vec_strings![
-        "-filter_complex",
-        "[0:v]null,zmq=b=tcp\\\\://'127.0.0.1\\:5555',drawtext=text=''"
-    ];
-    let params = vec_strings![
-        "-c:v",
-        "libx264",
-        "-flags",
-        "+global_header",
-        "-f",
-        "flv",
-        "rtmp://localhost/live/stream",
-        "-s",
-        "512x288",
-        "-c:v",
-        "libx264",
-        "-flags",
-        "+global_header",
-        "-f",
-        "flv",
-        "rtmp://localhost:1937/live/stream"
-    ];
-
-    let mut t1_params = enc_prefix.clone();
-    t1_params.append(&mut params.clone());
-    let cmd_two_outs =
-        prepare_output_cmd(enc_prefix.clone(), vec_strings![], params.clone(), "stream");
-
-    assert_eq!(cmd_two_outs, t1_params);
-
-    let mut test_cmd = enc_prefix.clone();
-    let mut test_params = params.clone();
-    let mut t2_filter = filter.clone();
-    t2_filter[1].push_str(",split=2[v_out1][v_out2]");
-    test_cmd.append(&mut t2_filter);
-
-    test_params.insert(0, "-map".to_string());
-    test_params.insert(1, "[v_out1]".to_string());
-    test_params.insert(2, "-map".to_string());
-    test_params.insert(3, "0:a".to_string());
-
-    test_params.insert(11, "-map".to_string());
-    test_params.insert(12, "[v_out2]".to_string());
-    test_params.insert(13, "-map".to_string());
-    test_params.insert(14, "0:a".to_string());
-
-    test_cmd.append(&mut test_params);
-    let cmd_two_outs_with_filter = prepare_output_cmd(enc_prefix, filter, params, "stream");
-
-    assert_eq!(cmd_two_outs_with_filter, test_cmd);
-}
